Corrected file size to be an integer.

+++ modules/content-creator/src/main/java/ccc/contentcreator/remoting/UpdateFileServlet.java
@@ -52,7 +52,7 @@ public class UpdateFileServlet extends CreatorServlet {
                     title.getString(),
                     description.getString(),
                     new MimeType(file.getContentType()),
-                    file.getSize(),
+                    (int) file.getSize(),
                     dataStream
                 );
             } finally {
